requirements.txt
streamlit
pandas
JPype1